+++ R/files.R
@@ -250,7 +250,7 @@ read_events_sleepedfx <- function(dir, update = TRUE){
 #' @return A dataframe of stages.
 #' @export
 read_events_compumedics <- function(
-    txt, startTime, labels = c("AWA" = 0, "N1" = 1, "N2" = 2, "N3" = 3, "REM" = 5)){
+    txt, startTime = as.POSIXlt("2000-01-01"), labels = c("AWA" = 0, "N1" = 1, "N2" = 2, "N3" = 3, "REM" = 5)){
   hypno <- utils::read.table(txt, stringsAsFactors = FALSE, col.names = "event")
   hypno$begin <- as.POSIXlt(startTime) + ((c(1:nrow(hypno))-1)*4)
   hypno$end <- hypno$begin+4