+++ NEWS.md
@@ -1,3 +1,13 @@
+# rsleep 1.0.12
+
+  * DOI: 10.5281/zenodo.10507974
+  * Added `detect_rem()` to detect rapid eye movements, implemented from Agarwal & Al, 2005.
+  * Added `detect_apneic_events()` to detect apneic events from oximetry signal, implemented from Jung & Al, 2018.
+  * Set default `startTime` to `2000-01-01` in `read_events_compumedics()`.
+  * Added new publications using the rsleep package
+  * `NULL` epochs are not removed anymore in the `epochs` function.
+  * Making sure the package build and test can run without an internet connection.
+
 # rsleep 1.0.11
 
   * Added `create_xts()` function to create `xts` objects from signal.